execute ARGV leftover rather than shell if present
Beginning with this commit, it is possible to do:

alip@karpov $ RUBY_VERSION=ruby-2.3 script/rvm.pl gemset default echo '$RUBY_VERSION'
ruby-2.3.3

which makes it convenient to select a command under a specific ruby version & gemset.

Leaving the minor version out makes for convenient usage and the ability
to run a command under a certain environment is good for Jenkins.